add new screen in project
