+++ lib/shop_app/constant/constant_screen.dart
@@ -75,3 +75,4 @@ void printAllText(String text) {
 // طيب هو احنا دلوقتي هنعرف توكن علشان اقدر اسنخدمه ي ال home lyout
 //ومنها اقدر احدد المستخدم الخاص بي
 String token = '';
+String uId = '';
