+++ lib/shop_app/screens/login/shop_login_screen.dart
@@ -154,6 +154,7 @@ class ShopLoginScreen extends StatelessWidget {
                                       password: passowrdController.text,
                                     );
                                   }
+                                  navigateTo(context, ShopLayout());
                                 },
                                 child: const Text(
                                   'LOGIN',
